Commit Code

+++ IBS/DataAccess/T15PoDetail.cs
@@ -7,7 +7,7 @@ public partial class T15PoDetail
 {
     public string CaseNo { get; set; } = null!;
 
-    public byte ItemSrno { get; set; }
+    public int ItemSrno { get; set; }
 
     public string? ItemDesc { get; set; }
